Suppression du dossier commandes et utils + Mise en place du EF!staff
